Create Dynamic programming 2
Quinton was given an array of integers and was asked to find the minimum subarray sum. A subarray is a contigous non-empty segment of an array.

Help Quinton solve this problem.

Input Description:
The first line contains an integer N denoting the number of citizens. The next line contains N space separated integers A[i] (1 <= i <= N)

Output Description:
Print a single line, the minimum subarray sum.